Replace SetVolumeButtonTest with snapshot test and remove unused sematics method
