+++ audio-ui/src/test/kotlin/com/google/android/horologist/audio/ui/components/actions/SetVolumeButtonTest.kt
@@ -1,0 +1,77 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@file:OptIn(ExperimentalHorologistPaparazziApi::class)
+
+package com.google.android.horologist.audio.ui.components.actions
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import com.google.android.horologist.audio.VolumeState
+import com.google.android.horologist.paparazzi.ExperimentalHorologistPaparazziApi
+import com.google.android.horologist.paparazzi.WearPaparazzi
+import org.junit.Rule
+import org.junit.Test
+
+class SetVolumeButtonTest {
+
+    @get:Rule
+    val paparazzi = WearPaparazzi()
+
+    @Test
+    fun givenCurrentVolumeIsNotMaxAndNotMin_thenIconIsVolumeDown() {
+        val currentVolume = 5
+
+        paparazzi.snapshot {
+            Box(modifier = Modifier.background(Color.Black), contentAlignment = Alignment.Center) {
+                SetVolumeButton(
+                    onVolumeClick = {},
+                    volumeState = VolumeState(current = currentVolume, max = 10)
+                )
+            }
+        }
+    }
+
+    @Test
+    fun givenCurrentVolumeIsMinimum_thenIconIsVolumeMute() {
+        val currentVolume = 0
+
+        paparazzi.snapshot {
+            Box(modifier = Modifier.background(Color.Black), contentAlignment = Alignment.Center) {
+                SetVolumeButton(
+                    onVolumeClick = {},
+                    volumeState = VolumeState(current = currentVolume, max = 10)
+                )
+            }
+        }
+    }
+
+    @Test
+    fun givenCurrentVolumeIsMaximum_thenIconIsVolumeUp() {
+        val currentVolume = 10
+        paparazzi.snapshot {
+            Box(modifier = Modifier.background(Color.Black), contentAlignment = Alignment.Center) {
+                SetVolumeButton(
+                    onVolumeClick = {},
+                    volumeState = VolumeState(current = currentVolume, max = currentVolume)
+                )
+            }
+        }
+    }
+}
