+++ audio-ui/src/test/snapshots/images/com.google.android.horologist.audio.ui.components.actions_SetVolumeButtonTest_givenCurrentVolumeIsMaximum_thenIconIsVolumeUp.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:87ee798fad3a9fc3b5a31dff8c76861e7b79b1c883ee05d61ed8d8d2f102cf4f
+size 28207
