+++ audio-ui/src/test/snapshots/images/com.google.android.horologist.audio.ui.components.actions_SetVolumeButtonTest_givenCurrentVolumeIsMinimum_thenIconIsVolumeMute.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:77883f6a8fecadcbb1c3d6bacea995abfa5e7531504cc4ec87e3d9e3525e512c
+size 23707
