+++ audio-ui/src/test/snapshots/images/com.google.android.horologist.audio.ui.components.actions_SetVolumeButtonTest_givenCurrentVolumeIsNotMaxAndNotMin_thenIconIsVolumeDown.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:515007ee2a4e937901bbb34417242818f36a0b864e083720ffb25ff47438b6d5
+size 24469
